+++ compiler/src/iree/compiler/Codegen/LLVMCPU/Passes.cpp
@@ -604,9 +604,8 @@ void addCPULinalgExtTileAndVectorizePipeline(
   addTileAndDistributePasses(funcPassManager);
   funcPassManager.addPass(
       createLLVMCPUTilePass(tilingConfig.getVectorCommonParallelLevel()));
-  // TODO: Should only apply decomposition here?
-  funcPassManager.addPass(
-      IREE::LinalgExt::createTileAndDecomposeAttentionPass());
+  funcPassManager.addPass(IREE::LinalgExt::createTileAttentionPass());
+  funcPassManager.addPass(IREE::LinalgExt::createDecomposeAttentionPass());
   funcPassManager.addPass(
       IREE::LinalgExt::createDecomposeWinogradTransformPass());
