+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/BUILD.bazel
@@ -32,12 +32,13 @@ iree_compiler_cc_library(
     srcs = [
         "ConvertConv2DToWinograd.cpp",
         "ConvertToLoops.cpp",
+        "DecomposeAttention.cpp",
         "DecomposeWinogradPass.cpp",
         "PadContractionToBlockSize.cpp",
         "PassDetail.h",
         "Passes.cpp",
         "SplitReduction.cpp",
-        "TileAndDecomposeAttention.cpp",
+        "TileAttention.cpp",
         "TilingInterfaceImpl.cpp",
     ],
     hdrs = [
