+++ compiler/src/iree/compiler/Dialect/LinalgExt/Transforms/Passes.h
@@ -46,13 +46,6 @@ createDecomposeWinogradTransformPass();
 std::unique_ptr<InterfacePass<mlir::FunctionOpInterface>>
 createConvertConv2DToWinogradPass();
 
-// Transform dialect version of tile and decompose attention wrapper.
-// The optional tile size specifies the step for the innermost for loop.
-void tileAndDecomposeAttention(IREE::LinalgExt::AttentionOp attnOp,
-                               SmallVectorImpl<Operation *> &ops,
-                               RewriterBase &rewriter, bool onlyTile = false,
-                               std::optional<uint64_t> tileSize = std::nullopt);
-
 IREE::LinalgExt::AttentionOp
 tileAttention(IREE::LinalgExt::AttentionOp attnOp,
               SmallVectorImpl<Operation *> &ops, RewriterBase &rewriter,
@@ -63,9 +56,11 @@ void decomposeTiledAttention(IREE::LinalgExt::AttentionOp tiledAttnOp,
                              RewriterBase &rewriter,
                              std::optional<uint64_t> tileSize = std::nullopt);
 
-// Creates a pass to convert the attention op into a sequence of
-// linalg ops.
-std::unique_ptr<Pass> createTileAndDecomposeAttentionPass();
+// Creates a pass to tile the attention op along the reduction dim.
+std::unique_ptr<Pass> createTileAttentionPass();
+
+// Creates a pass to convert the attention op into a sequence of linalg ops.
+std::unique_ptr<Pass> createDecomposeAttentionPass();
 
 //===---------------------------------------------------------------------===//
 // Codegen Strategy passes that are moved into IREE.
